% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_range.R
\name{age_range}
\alias{age_range}
\title{Age Range}
\usage{
age_range(age)
}
\arguments{
\item{age}{numeric, the age the the output will be relative to}
}
\value{
The ages that are the appropriate age range for the age entered
}
\description{
Prints the age range for a specific age, when following the "Half-Your-Age-Plus-Seven" Rule
}
\examples{
age_range(28)

}
