% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_couple.R
\name{check_couple}
\alias{check_couple}
\title{Checks Whether Couple Meets Rule}
\usage{
check_couple(age1, age2)
}
\arguments{
\item{age1}{numeric, the age of the first member of the couple}

\item{age2}{numeric, the age of the second member of the couple}
}
\value{
The verdict about the couple's ages using the "Half-Your-Age-Plus-Seven" Rule
}
\description{
Using the "Half-Your-Age-Plus-Seven" Rule, the function returns the "verdict" given two ages for the members of a couple
}
\examples{
check_couple(70, 30)

check_couple(-10, 19)

check_couple(50, 80)


}
