% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upper_limit.R
\name{upper_limit}
\alias{upper_limit}
\title{Returns Upper Age Limit of Age Range}
\usage{
upper_limit(age)
}
\arguments{
\item{age}{numeric, the age the the output will be relative to}
}
\value{
The age that is the upper limit for the age entered
}
\description{
Prints the upper age limit for a specific age, when following the "Half-Your-Age-Plus-Seven" Rule
}
\examples{
upper_limit(30)

upper_limit(15)

upper_limit(57)

}
