% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ages_firstok.R
\name{ages_firstok}
\alias{ages_firstok}
\title{First Okay}
\usage{
ages_firstok(birthdate1, birthdate2)
}
\arguments{
\item{birthdate1}{date, the date when person 1 was born}

\item{birthdate2}{date, the date when person 2 was born}
}
\value{
The ages that are the appropriate age range for the age entered
}
\description{
Prints the ages when two people of the entered birth dates will have an appropriate age gap, given the "Half-Your-Age-Plus-7" Rule.
}
\examples{
ages_firstokay(1970-03-19, 2000-07-09)

}
