% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower_limit.R
\name{lower_limit}
\alias{lower_limit}
\title{Lower Age Limit}
\usage{
lower_limit(age)
}
\arguments{
\item{age}{numeric, the age the the output will be relative to}
}
\value{
The age that is the lower limit for the age entered
}
\description{
Prints the lower age limit for a specific age, when following the "Half-Your-Age-Plus-Seven" Rule
}
\examples{
lower_limit(30)

}
